requirements.txt
Django==5.0.6
gunicorn==22.0.0
psycopg[binary]==3.2.9
whitenoise==6.7.0
python-dotenv==1.0.1
requests==2.32.3
django-parler==2.3
django-tailwind==4.2.0
django-browser-reload==1.18.0
django-widget-tweaks==1.5.0
qrcode==7.4.2
Pillow==10.4.0
WeasyPrint==61.2
psycopg2-binary==2.9.9
reportlab==4.2.0
PyPDF2
xhtml2pdf




